stream/requirements.txt
gqlalchemy==1.0.5
kafka-python==2.0.2
pandas
